EnumExtension class to new file
